refactor: Point -> Position으로 변경
