+++ src/main/kotlin/omok/domain/OmokGrid.kt
@@ -1,6 +1,5 @@
 package omok.domain
 
-import rule.wrapper.point.Point
 
 fun List<MutableList<StoneState>>.deepCopy(): List<MutableList<StoneState>> = map { it.toMutableList() }.toList()
 
@@ -9,15 +8,15 @@ class OmokGrid(val width: Int = DEFAULT_SIZE, val height: Int = DEFAULT_SIZE) {
     val board: List<MutableList<StoneState>>
         get() = _board.deepCopy()
 
-    fun canPlace(point: Point) {
-        if (_board[point.row][point.col] != StoneState.BLANK) throw IllegalStateException(ERROR_STONE_ALREADY_PUT)
+    fun canPlace(point: Position) {
+        if (_board[point.y][point.x] != StoneState.BLANK) throw IllegalStateException(ERROR_STONE_ALREADY_PUT)
     }
 
     fun putStone(
-        point: Point,
+        point: Position,
         state: StoneState,
     ) {
-        _board[point.row][point.col] = state
+        _board[point.y][point.x] = state
     }
 
     fun isFull(): Boolean {
