+++ src/main/kotlin/omok/domain/Position.kt
