+++ src/main/kotlin/omok/domain/StoneState.kt
@@ -1,22 +1,8 @@
 package omok.domain
 
-import omok.domain.player.BlackPlayer
-import omok.domain.player.Player
-import omok.domain.player.WhitePlayer
-
 enum class StoneState {
     BLACK,
     WHITE,
     BLANK,
     ;
-
-    companion object {
-        fun getColor(player: Player): StoneState {
-            return when (player) {
-                is BlackPlayer -> BLACK
-                is WhitePlayer -> WHITE
-                else -> throw IllegalStateException()
-            }
-        }
-    }
 }
