+++ src/main/kotlin/omok/domain/lib/FourFourRule.kt
