+++ src/main/kotlin/omok/domain/lib/OmokRule.kt
