+++ src/main/kotlin/omok/domain/lib/ThreeThreeRule.kt
