+++ src/main/kotlin/omok/domain/player/Player.kt
@@ -1,7 +1,6 @@
 package omok.domain.player
 
 import rule.OmokRule
-import rule.type.Violation
 import rule.wrapper.point.Point
 
 abstract class Player {
@@ -23,20 +22,20 @@ abstract class Player {
         startPoint: Point,
     )
 
-    companion object {
-        private const val WIN_STANDARD = 5
-
-        private const val ERROR_DOUBLE_THREE = "3x3 위치에 놓을 수 없습니다"
-        private const val ERROR_DOUBLE_FOUR = "4x4 위치에 놓을 수 없습니다"
-        private const val ERROR_OVER_LINE = "장목 위치에 놓을 수 없습니다"
-
-        fun dealViolation(violation: Violation) {
-            when (violation) {
-                Violation.DOUBLE_THREE -> throw IllegalStateException(ERROR_DOUBLE_THREE)
-                Violation.DOUBLE_FOUR -> throw IllegalStateException(ERROR_DOUBLE_FOUR)
-                Violation.OVERLINE -> throw IllegalStateException(ERROR_OVER_LINE)
-                Violation.NONE -> {}
-            }
-        }
-    }
+//    companion object {
+//        private const val WIN_STANDARD = 5
+//
+//        private const val ERROR_DOUBLE_THREE = "3x3 위치에 놓을 수 없습니다"
+//        private const val ERROR_DOUBLE_FOUR = "4x4 위치에 놓을 수 없습니다"
+//        private const val ERROR_OVER_LINE = "장목 위치에 놓을 수 없습니다"
+//
+//        fun dealViolation(violation: Violation) {
+//            when (violation) {
+//                Violation.DOUBLE_THREE -> throw IllegalStateException(ERROR_DOUBLE_THREE)
+//                Violation.DOUBLE_FOUR -> throw IllegalStateException(ERROR_DOUBLE_FOUR)
+//                Violation.OVERLINE -> throw IllegalStateException(ERROR_OVER_LINE)
+//                Violation.NONE -> {}
+//            }
+//        }
+//    }
 }
