+++ src/main/kotlin/omok/view/InputView.kt
@@ -1,26 +1,16 @@
 package omok.view
 
 import omok.domain.OmokGrid
-import omok.domain.player.BlackPlayer
+import omok.domain.Position
 import omok.domain.player.Player
-import omok.domain.player.WhitePlayer
-import rule.wrapper.point.Point
-
-private fun Player.getDisplayColor(): String {
-    return when (this) {
-        is BlackPlayer -> "흑"
-        is WhitePlayer -> "백"
-        else -> throw IllegalArgumentException()
-    }
-}
 
 class InputView {
     fun getPoint(
         player: Player,
-        latestPoint: Point?,
+        latestPoint: Position?,
         grid: OmokGrid,
-    ): Point {
-        print(MESSAGE_TURN.format(player.getDisplayColor()))
+    ): Position {
+//        print(MESSAGE_TURN.format(player.getDisplayColor()))
         if (latestPoint != null) print(MESSAGE_LATEST_POSITION.format(convertToString(latestPoint)))
         print(MESSAGE_POSITION_GUIDE)
         val rawInput = readln().trim()
@@ -33,15 +23,15 @@ class InputView {
         private const val MESSAGE_LATEST_POSITION: String = "(마지막 돌의 위치: %s)"
         private const val MESSAGE_POSITION_GUIDE: String = "\n위치를 입력하세요: "
 
-        private fun convertToString(point: Point): String {
-            val letter = 'A' + point.col - 1
-            return letter + (point.row).toString()
+        private fun convertToString(point: Position): String {
+            val letter = 'A' + point.x - 1
+            return letter + (point.y).toString()
         }
 
         private fun parsingInput(
             rawInput: String,
             grid: OmokGrid,
-        ): Point? {
+        ): Position? {
             if (rawInput.isEmpty()) return null
 
             val rawRow = rawInput.substring(1)
@@ -49,7 +39,7 @@ class InputView {
 
             val row = validateRow(rawRow, grid.height) ?: return null
             val col = validateCol(rawCol, grid.width) ?: return null
-            return Point(row, col)
+            return Position(row, col)
         }
 
         private fun validateRow(
